refactor: layout check
